Extract config TUI helpers into dedicated module
- Move config display and edit UI helpers out of `src/cli/config.rs`
- Add `Features::has_configured_overrides()` for provider filtering
