+++ src/cli/ui/mod.rs
@@ -1,4 +1,5 @@
 pub(crate) mod common;
+pub(crate) mod config;
 pub(crate) mod deploy;
 pub(crate) mod dry_run;
 pub(crate) mod init;
